requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
openai
pydantic
pydantic-settings
python-multipart
pypdf
python-docx
openpyxl
pytest
# Google Drive integration
google-api-python-client
google-auth
